+++ src/main/java/design/Model/Food/Meal.java
@@ -25,8 +25,8 @@ public class Meal implements Food{
     }
 
     @Override
-    public Integer getFat() {
-        Integer fat = 0;
+    public Double getFat() {
+        Double fat = 0.0;
         for (Recipe recipe : recipes){
             fat += recipe.getFat();
         }
@@ -34,8 +34,8 @@ public class Meal implements Food{
     }
 
     @Override
-    public Integer getProtein() {
-        Integer protein = 0;
+    public Double getProtein() {
+        Double protein = 0.0;
         for (Recipe recipe : recipes){
             protein += recipe.getProtein();
         }
@@ -43,8 +43,8 @@ public class Meal implements Food{
     }
 
     @Override
-    public Integer getFiber() {
-        Integer fiber = 0;
+    public Double getFiber() {
+        Double fiber = 0.0;
         for (Recipe recipe : recipes){
             fiber += recipe.getFiber();
         }
@@ -52,8 +52,8 @@ public class Meal implements Food{
     }
 
     @Override
-    public Integer getCarbs() {
-        Integer carbs = 0;
+    public Double getCarbs() {
+        Double carbs = 0.0;
         for (Recipe recipe : recipes){
             carbs += recipe.getCarbs();
         }
